In an SSH session check forwarded agent first
When in an SSH session, try to connect to a forwarded agent first. If
that fails, only then check if an SSH agent is already running on the
machine.

This fixes a situation where a .ssh/environment file left behind from
a dead agent is checked first, ignoring the forwarded agent and
starting a new agent on the machine.